[core] use unique IDs for DrawScopes

+++ src/mbgl/programs/program_parameters.cpp
@@ -34,7 +34,7 @@ optional<std::string> ProgramParameters::cachePath(const char* name) const {
         result += "/com.mapbox.gl.shader.";
         result += name;
         result += '.';
-        result += util::toHex(std::hash<std::string>()(defines));
+        result += util::toHex(static_cast<uint64_t>(std::hash<std::string>()(defines)));
         result += ".pbf";
         return result;
     }
